# repo/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Philatelic Museum Operations Suite API
  version: 1.0.0
  description: |
    Public HTTP API for museum operations.

    Authentication model:
    - Session cookie `museum_sid` (set on `POST /auth/login`)
    - `X-CSRF-Token` required on mutating requests (`POST|PUT|PATCH|DELETE`) except `POST /auth/login`
    - `X-Step-Up-Token` required on sensitive actions after `POST /auth/step-up`

    Step-up actions:
    - `GRAPH_PUBLISH`
    - `ROUTE_RULE_CHANGE`
    - `JOB_APPROVE`
    - `JOB_APPEAL_DECIDE`
    - `EXPORT_CREATE`
    - `ADMIN_CONFIG_UPDATE`
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - SessionCookie: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users (excluding password hashes)
      description: Requires permission `USERS_ADMIN` (Administrator).
      responses:
        '200': { description: User list }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      summary: Create user
      description: Requires permission `USERS_ADMIN` (Administrator).
      security:
        - SessionCookie: []
          CsrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, roles]
              properties:
                username: { type: string }
                password: { type: string }
                roles:
                  type: array
                  items: { $ref: '#/components/schemas/Role' }
      responses:
        '201': { description: User created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/{userId}:
    patch:
      tags: [Users]
      summary: Update user roles/status
      description: Requires permission `USERS_ADMIN` (Administrator).
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: User updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /venues:
    post:
      tags: [Venues & Routes]
      summary: Create venue
      description: Requires `VENUE_MANAGE` (Administrator, Exhibit Manager).
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Venue created }
  /venues/{venueId}/halls:
    post:
      tags: [Venues & Routes]
      summary: Create hall
      description: Requires `VENUE_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201': { description: Hall created }
  /halls/{hallId}/zones:
    post:
      tags: [Venues & Routes]
      summary: Create zone
      description: Requires `VENUE_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: hallId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201': { description: Zone created }
  /zones/{zoneId}/display-cases:
    post:
      tags: [Venues & Routes]
      summary: Create display case
      description: Requires `VENUE_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: zoneId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201': { description: Display case created }
  /routes:
    post:
      tags: [Venues & Routes]
      summary: Create route definition
      description: Requires `VENUE_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Route created }
  /routes/{routeId}/segments:
    post:
      tags: [Venues & Routes]
      summary: Add route segment
      description: Requires `VENUE_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: routeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '201': { description: Segment created }
        '400': { $ref: '#/components/responses/ValidationError' }
  /routes/{routeId}:
    patch:
      tags: [Venues & Routes]
      summary: Update route rules
      description: Requires `ROUTE_RULE_CHANGE` and step-up action `ROUTE_RULE_CHANGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
          StepUpHeader: []
      parameters:
        - name: routeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Route updated }
  /routes/{routeId}/itineraries:
    post:
      tags: [Venues & Routes]
      summary: Generate itinerary from route segments
      description: Requires `VENUE_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: routeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            example:
              accessibilityMode: true
              branchSelections:
                - fromCaseId: 65f0f7d1d5f5f8612f06c101
                  toCaseId: 65f0f7d1d5f5f8612f06c102
      responses:
        '201':
          description: Itinerary generated
          content:
            application/json:
              example:
                data:
                  itineraryId: iti_8a6c9e11
                  estimatedWalkMinutes: 7.1
                  printable:
                    routeId: rte_10f7ab22
                    accessibilityMode: true
                    steps:
                      - step: 1
                        segmentType: REQUIRED_NEXT
                        fromCaseId: '65f0f7d1d5f5f8612f06c101'
                        toCaseId: '65f0f7d1d5f5f8612f06c102'
                        dwellMinutes: 4
                        distanceMeters: 40
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: museum_sid
      description: Session cookie set by `POST /auth/login`.
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on mutating requests except `POST /auth/login`.
    StepUpHeader:
      type: apiKey
      in: header
      name: X-Step-Up-Token
      description: One-time action-bound step-up token for sensitive operations.
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    Role:
      type: string
      enum:
        - Administrator
        - Curator
        - Exhibit Manager
        - Program Coordinator
        - Employer
        - Reviewer
        - Auditor
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              oneOf:
                - type: array
                  items: { type: object }
                - type: object
            requestId: { type: string }
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: Insufficient role permission or object-level access denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: State conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...